requirements.txt
psycopg2-binary==2.9.7
requests==2.31.0
gunicorn==21.2.0
python-dotenv==1.0.0
supabase==2.18.1
fastapi==0.115.0
uvicorn==0.32.0
